pyproject.toml
[project]
name = "tctk"
version = "0.1"
description = ""
readme = "README.md"
authors = [{name = ""}]
license = "Apache-2.0"
requires-python = ">=3.8"
classifiers = []
dependencies = [
    "asyncclick",
    "elasticsearch>=8.12.0",
    "polars>=1.8.2",
    "pydash>=8.0.5",
    "pytest>=8.3.5",
    "python-box[all]>=7.2.0",
    "pyyaml",
    "twitchAPI",
    "SQLAlchemy>=2.0",
    # pandas: pick a compatible version for Python < 3.9
    "pandas>=2.0,<2.2; python_version < '3.9'",
    "pandas>=2.2; python_version >= '3.9'",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
tctk = "tctk.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
